Updated Project. Rn-starter descontinued. Wrong directory.
Created customised component with an image and a header for all screens.
Created customised component with image and title for TopDishes screen.
Learn and implemented Linking method on TopDishes screen so user may click on button to redirect them to a google search about desired dish so they may learn more on google.
Added images to nearly every screen left to fill with content.
3 out of 7 Screens completed.